added highlighting for demotext, cleaned code
