+++ src/test/kotlin/org/openpolicyagent/ideaplugin/ide/colors/RegoColorSettingsPageTest.kt
@@ -5,8 +5,10 @@
 
 package org.openpolicyagent.ideaplugin.ide.colors
 
+import com.intellij.openapi.editor.colors.TextAttributesKey
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.Test
+import org.openpolicyagent.ideaplugin.ide.extensions.RegoHighlighterAnnotator
 import org.openpolicyagent.ideaplugin.ide.highlight.RegoHighlighter
 
 class RegoColorSettingsPageTest {
@@ -20,10 +22,13 @@ class RegoColorSettingsPageTest {
     @Test
     fun `token defined in the color settings page should be defined in the highlighter`() {
         val colorSettingsAttributes = RegoColorSettingsPage().attributeDescriptors.map { it.key }
+        val syntaxHighlighterColors : List<TextAttributesKey> = RegoHighlighter().tokenToColorMap.values.distinct()
+        val highlighterAnnotatorColors = RegoHighlighterAnnotator().usedColors
+        val allUsedColors : List<TextAttributesKey> = syntaxHighlighterColors + highlighterAnnotatorColors
 
         assertThat(colorSettingsAttributes)
-            .describedAs("Some token are defined in the color settings page but not in Highlighter. That mean user can choose color for this token but ide will never render it")
-            .containsExactlyInAnyOrderElementsOf(RegoHighlighter().tokenToColorMap.values.distinct())
+            .describedAs("Some token are defined in the color settings page but not in Highlighter or Annotator. That mean user can choose color for this token but ide will never render it")
+            .containsExactlyInAnyOrderElementsOf(allUsedColors)
 
 
     }
